Abstract out ControlPlane interface
Eventually we will have a new implementation of `ControlPlane`, different from the in-memory one.
