+++ storage/inkless/src/main/java/io/aiven/inkless/common/SharedState.java
@@ -16,7 +16,7 @@ public record SharedState(
 ) {
 
     public static SharedState initialize(Time time, InklessConfig config, MetadataView metadata) {
-        return new SharedState(time, config, metadata, new ControlPlane(metadata), config.storage());
+        return new SharedState(time, config, metadata, ControlPlane.init(config, metadata), config.storage());
     }
 
 }
