+++ storage/inkless/src/main/java/io/aiven/inkless/control_plane/ControlPlane.java
@@ -1,118 +1,28 @@
-// Copyright (c) 2024 Aiven, Helsinki, Finland. https://aiven.io/
 package io.aiven.inkless.control_plane;
 
-import java.util.ArrayList;
-import java.util.HashMap;
+import java.lang.reflect.Constructor;
+import java.lang.reflect.InvocationTargetException;
 import java.util.List;
-import java.util.Map;
-import java.util.NavigableSet;
-import java.util.Set;
-import java.util.TreeMap;
-
-import org.apache.kafka.common.TopicIdPartition;
-import org.apache.kafka.common.TopicPartition;
-import org.apache.kafka.common.Uuid;
 
 import io.aiven.inkless.common.ObjectKey;
+import io.aiven.inkless.config.InklessConfig;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+public interface ControlPlane {
+    List<CommitBatchResponse> commitFile(ObjectKey objectKey,
+                                         List<CommitBatchRequest> batches);
 
-public class ControlPlane {
-    private static final Logger logger = LoggerFactory.getLogger(ControlPlane.class);
+    List<FindBatchResponse> findBatches(List<FindBatchRequest> findBatchRequests,
+                                        boolean minOneMessage,
+                                        int fetchMaxBytes);
 
-    private final MetadataView metadataView;
-
-    private final Map<TopicIdPartition, LogInfo> logs = new HashMap<>();
-    private final HashMap<TopicIdPartition, TreeMap<Long, BatchInfo>> batches = new HashMap<>();
-
-    public ControlPlane(final MetadataView metadataView) {
-        this.metadataView = metadataView;
-    }
-
-    public synchronized List<CommitBatchResponse> commitFile(final ObjectKey objectKey,
-                                                             final List<CommitBatchRequest> batches) {
-        final List<CommitBatchResponse> responses = new ArrayList<>();
-
-        for (final CommitBatchRequest request : batches) {
-            final String topicName = request.topicPartition().topic();
-            final Uuid topicId = metadataView.getTopicId(topicName);
-            final Set<TopicPartition> partitions = metadataView.getTopicPartitions(topicName);
-            if (topicId == Uuid.ZERO_UUID
-                || !partitions.contains(request.topicPartition())) {
-                responses.add(CommitBatchResponse.unknownTopicOrPartition());
-            } else {
-                final TopicIdPartition topicIdPartition = new TopicIdPartition(topicId, request.topicPartition());
-                final LogInfo logInfo = logs.computeIfAbsent(topicIdPartition, ignore -> new LogInfo());
-                final long firstOffset = logInfo.highWatermark;
-                logInfo.highWatermark += request.numberOfRecords();
-                final long lastOffset = logInfo.highWatermark - 1;
-                // TODO: also compute append timestamp
-                final BatchInfo batchToStore = new BatchInfo(objectKey, request.byteOffset(), request.size(), firstOffset, request.numberOfRecords());
-                this.batches
-                    .computeIfAbsent(topicIdPartition, ignore -> new TreeMap<>())
-                    .put(lastOffset, batchToStore);
-                responses.add(CommitBatchResponse.success(firstOffset, logInfo.logStartOffset));
-            }
+    static ControlPlane init(final InklessConfig config, final MetadataView metadata) {
+        final Class<ControlPlane> controlPlaneClass = config.controlPlaneClass();
+        try {
+            final Constructor<ControlPlane> ctor = controlPlaneClass.getConstructor(MetadataView.class);
+            return ctor.newInstance(metadata);
+        } catch (final NoSuchMethodException | InstantiationException | IllegalAccessException |
+                       InvocationTargetException e) {
+            throw new RuntimeException(e);
         }
-
-        return responses;
-    }
-
-    public synchronized List<FindBatchResponse> findBatches(final List<FindBatchRequest> findBatchRequests,
-                                                           final boolean minOneMessage,
-                                                           final int fetchMaxBytes) {
-        final List<FindBatchResponse> result = new ArrayList<>();
-
-        for (final FindBatchRequest request : findBatchRequests) {
-            final String topicName = request.topicIdPartition().topic();
-            final Uuid topicId = metadataView.getTopicId(topicName);
-            final Set<TopicPartition> partitions = metadataView.getTopicPartitions(topicName);
-            if (!topicId.equals(request.topicIdPartition().topicId())
-                || !partitions.contains(request.topicIdPartition().topicPartition())) {
-                result.add(FindBatchResponse.unknownTopicOrPartition());
-            } else {
-                final LogInfo logInfo = logs.computeIfAbsent(request.topicIdPartition(), ignore -> new LogInfo());
-                if (request.offset() < 0) {
-                    logger.debug("Invalid offset {} for {}", request.offset(), request.topicIdPartition());
-                    result.add(FindBatchResponse.offsetOutOfRange(
-                        logInfo.logStartOffset, logInfo.highWatermark));
-                } else {
-                    if (request.offset() >= logInfo.highWatermark) {
-                        result.add(FindBatchResponse.offsetOutOfRange(
-                            logInfo.logStartOffset, logInfo.highWatermark));
-                    } else {
-                        final TreeMap<Long, BatchInfo> coordinates = this.batches.get(request.topicIdPartition());
-                        if (coordinates != null) {
-                            List<BatchInfo> batches = new ArrayList<>();
-                            long totalSize = 0;
-                            for (Long batchOffset : coordinates.navigableKeySet().tailSet(request.offset())) {
-                                BatchInfo batch = coordinates.get(batchOffset);
-                                batches.add(batch);
-                                totalSize += batch.size();
-                                if (totalSize > fetchMaxBytes) {
-                                    break;
-                                }
-                            }
-                            result.add(FindBatchResponse.success(
-                                batches, logInfo.logStartOffset, logInfo.highWatermark));
-                        } else {
-                            logger.error("Batch coordinates not found for {}: high watermark={}, requested offset={}",
-                                request.topicIdPartition(),
-                                logInfo.highWatermark,
-                                request.offset());
-                            result.add(FindBatchResponse.unknownServerError());
-                        }
-                    }
-                }
-            }
-        }
-
-        return result;
-    }
-
-    private static class LogInfo {
-        long logStartOffset = 0;
-        long highWatermark = 0;
     }
 }
